[REFACTOR & STYLES]: Use nested layouts
The navigation sidebar will not be unique to the sets page, so
I extracted it to its own layout. Nested layouts are necessary
because every page should have the header, but only some pages
should show the sidebar.

The new syntax renders the `set` template in the `nav_on_left`
template, which is in turn embedded in the default `layout` template.